+++ CHANGELOG.md
@@ -17,12 +17,17 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds ability to change the color of an SVG icon with classes such as: `ui-black` (default), `ui-white`, `brand-primary` (red), `section-research-secondary` (turquoise).
 - Adds `helperText` and `errorText` props to `Checkbox`
 - Adds `helperText` and `errorText` props to `Radio`
+- Adds `errorText`, `helperText`, `labelText`, `required` and `showLabel` props to `Select`
 
 ### Changes
 
 - Updates content in Style Guide `Typography`
 - Updates content in Style Guide `Breakpoints`
 - Changes `Link` story from `.tsx` to `.mdx`.
+
+### Deprecates
+
+- Deprecates `ariaLabel`, `helperTextId`, `isRequired` and `labelId` props in `Select`
 
 ## 0.23.0